+++ play-simple-webdav/src/main/java/nl/idgis/dav/model/DefaultResourceDescription.java
@@ -1,4 +1,4 @@
-package model.dav;
+package nl.idgis.dav.model;
 
 public class DefaultResourceDescription implements ResourceDescription {
