+++ publisher-metadata/app/controllers/ServiceMetadata.java
@@ -12,13 +12,13 @@ import com.mysema.query.types.Predicate;
 import com.mysema.query.types.QTuple;
 import com.typesafe.config.Config;
 
-import model.dav.Resource;
-import model.dav.ResourceDescription;
-import model.dav.ResourceProperties;
+import nl.idgis.dav.model.Resource;
+import nl.idgis.dav.model.ResourceDescription;
+import nl.idgis.dav.model.ResourceProperties;
 
-import model.dav.DefaultResource;
-import model.dav.DefaultResourceDescription;
-import model.dav.DefaultResourceProperties;
+import nl.idgis.dav.model.DefaultResource;
+import nl.idgis.dav.model.DefaultResourceDescription;
+import nl.idgis.dav.model.DefaultResourceProperties;
 
 import nl.idgis.publisher.metadata.MetadataDocument;
 import nl.idgis.publisher.metadata.MetadataDocumentFactory;
@@ -30,8 +30,6 @@ import play.api.routing.Router;
 import play.libs.Json;
 import play.mvc.Controller;
 import play.mvc.Result;
-
-import router.dav.SimpleWebDAV;
 
 import util.InetFilter;
 import util.MetadataConfig;
